requirements.txt
alpaca==1.0.0
alpaca_backtrader_api==0.15.0
backtrader==1.9.78.123
pandas==1.5.3
plotly==5.13.0
python-dotenv==1.0.0
